Move some types from tensorzero-core to leaf crates
This is progress towards removing the `tensorzero-core` dependency
from `ts-executor-pool`
