+++ crates/tensorzero-types-providers/Cargo.toml
@@ -10,6 +10,7 @@ serde.workspace = true
 serde_json.workspace = true
 serde_path_to_error.workspace = true
 tensorzero-derive = { path = "../tensorzero-derive" }
+tensorzero-types = { path = "../tensorzero-types" }
 
 [lints]
 workspace = true
