+++ crates/tensorzero-types/src/lib.rs
@@ -9,6 +9,7 @@ pub mod error;
 pub mod file;
 pub mod inference_filters;
 pub mod inference_params;
+pub mod inference_response;
 pub mod message;
 pub mod rate_limiting_types;
 pub mod role;
@@ -38,6 +39,10 @@ pub use file::{
     Base64File, Base64FileMetadata, Detail, File, ObjectStorageError, ObjectStorageFile,
     ObjectStoragePointer, UrlFile,
 };
+pub use inference_response::{
+    ChatInferenceResponse, ContentBlockChatOutput, InferenceResponse, JsonInferenceOutput,
+    JsonInferenceResponse,
+};
 pub use message::{Input, InputMessage, InputMessageContent, TextKind};
 pub use role::{
     ASSISTANT_TEXT_TEMPLATE_VAR, Role, SYSTEM_TEXT_TEMPLATE_VAR, USER_TEXT_TEMPLATE_VAR,
@@ -50,7 +55,10 @@ pub use tool::{InferenceResponseToolCall, ToolCall, ToolCallWrapper, ToolChoice,
 pub use tool_error::ToolError;
 pub use tool_failure::{NonControlToolError, ToolFailure};
 pub use tool_handle::ToolHandle;
-pub use usage::{ApiType, RawResponseEntry};
+pub use usage::{
+    ApiType, FinishReason, OpenAIPromptTokensDetails, OpenAIUsage, RawResponseEntry, RawUsageEntry,
+    Usage, raw_usage_entries_from_value,
+};
 use uuid::Uuid;
 
 #[cfg_attr(feature = "ts-bindings", derive(ts_rs::TS))]
